requirements.txt
python-telegram-bot
flask
requests
aiohttp
firebase_admin